factor out polls to posts module
